# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Petrobras File Transfer API
  description: |
    API RESTful para o sistema de transferencia segura de arquivos da Petrobras.
    
    ## Autenticacao
    
    A API suporta dois metodos de autenticacao:
    
    1. **Bearer Token (JWT)** - Para usuarios internos autenticados via Entra ID ou login local
    2. **Cookie de Sessao** - Alternativa ao Bearer token para aplicacoes web
    
    ## Fluxo de Compartilhamento
    
    1. Usuario interno faz upload dos arquivos
    2. Cria um compartilhamento com email do destinatario externo
    3. Supervisor aprova ou rejeita a solicitacao
    4. Destinatario recebe email com link e codigo OTP
    5. Destinatario acessa o portal, valida OTP e baixa os arquivos
    
  version: 2.0.0
  contact:
    name: Suporte Petrobras
    email: suporte@petrobras.com.br
  license:
    name: Proprietary
    url: https://petrobras.com.br
servers:
  - url: http://localhost:8000/api/v1
    description: Servidor de desenvolvimento local
  - url: https://api.petrobras.com.br/api/v1
    description: Servidor de producao
...
paths:
...
  /users/me:
    get:
      tags: [Users]
      summary: Obter perfil do usuario autenticado
      description: Retorna dados do usuario logado
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Dados do usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'
    put:
      tags: [Users]
      summary: Atualizar perfil
      description: Atualiza dados do perfil do usuario autenticado
      operationId: updateCurrentUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Perfil atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
  /users:
    get:
      tags: [Users]
      summary: Listar usuarios
      description: Lista todos os usuarios (admin)
      operationId: listUsers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de usuarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      tags: [Users]
      summary: Criar usuario
      description: Cria novo usuario (admin)
      operationId: createUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: Usuario criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/{userId}:
    get:
      tags: [Users]
      summary: Obter usuario por ID
      operationId: getUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/userId'
      responses:
        '200':
          description: Dados do usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Users]
      summary: Atualizar usuario
      operationId: updateUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/userId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Usuario atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

  # ============================================================
  # FILES - Arquivos
  # ============================================================
  /files:
    get:
      tags: [Files]
      summary: Listar arquivos
      description: Lista arquivos do usuario com paginacao
      operationId: listFiles
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: area_id
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [active, deleted]
      responses:
        '200':
          description: Lista de arquivos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileListResponse'
  /files/upload:
    post:
      tags: [Files]
      summary: Upload de arquivo
      description: Faz upload de arquivo via multipart/form-data
      operationId: uploadFile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                area_id:
                  type: integer
                description:
                  type: string
      responses:
        '201':
          description: Arquivo enviado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/File'
        '413':
          description: Arquivo muito grande
  /files/{fileId}:
    get:
      tags: [Files]
      summary: Obter detalhes do arquivo
      operationId: getFile
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/fileId'
      responses:
        '200':
          description: Detalhes do arquivo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/File'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Files]
      summary: Excluir arquivo
      operationId: deleteFile
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/fileId'
      responses:
        '200':
          description: Arquivo excluido
        '404':
          $ref: '#/components/responses/NotFound'
...
  /shares:
    get:
      tags: [Shares]
      summary: Listar compartilhamentos
      description: Lista compartilhamentos do usuario com paginacao
      operationId: listShares
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [pendente, ativo, aprovado, rejeitado, concluido, expirado, cancelado]
      responses:
        '200':
          description: Lista de compartilhamentos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShareListResponse'
    post:
      tags: [Shares]
      summary: Criar compartilhamento
      description: Cria novo compartilhamento de arquivos
      operationId: createShare
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateShareRequest'
      responses:
        '201':
          description: Compartilhamento criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Share'
  /shares/{shareId}:
    get:
      tags: [Shares]
      summary: Obter detalhes do compartilhamento
      operationId: getShare
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/shareId'
      responses:
        '200':
          description: Detalhes do compartilhamento
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShareDetails'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Shares]
      summary: Excluir compartilhamento
      operationId: deleteShare
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/shareId'
      responses:
        '200':
          description: Compartilhamento excluido
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT para usuarios internos
    cookieAuth:
      type: apiKey
      in: cookie
      name: app_session
      description: Cookie de sessao alternativo
    externalSession:
      type: apiKey
      in: cookie
      name: external_session
      description: Cookie de sessao para usuarios externos
  parameters:
    userId:
      name: userId
      in: path
      required: true
      schema:
        type: integer
      description: ID do usuario
    fileId:
      name: fileId
      in: path
      required: true
      schema:
        type: integer
      description: ID do arquivo
    shareId:
      name: shareId
      in: path
      required: true
      schema:
        type: integer
      description: ID do compartilhamento
  responses:
    Unauthorized:
      description: Nao autenticado
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
                example: Nao autenticado
    NotFound:
      description: Recurso nao encontrado
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
                example: Recurso nao encontrado
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          format: email
        type:
          type: string
          enum: [externo, internal, supervisor]
        status:
          type: boolean
        created_at:
          type: string
          format: date-time
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
        department:
          type: string
        job_title:
          type: string
        phone:
          type: string
        employee_id:
          type: string
        photo_url:
          type: string
        created_at:
          type: string
          format: date-time
        last_login:
          type: string
          format: date-time
        manager:
          type: object
          nullable: true
          properties:
            id:
              type: integer
            name:
              type: string
            email:
              type: string
    CreateUserRequest:
      type: object
      required: [name, email, type]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        type:
          type: string
          enum: [externo, internal, supervisor]
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        type:
          type: string
        status:
          type: boolean
    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        department:
          type: string
        job_title:
          type: string

    # File Schemas
    File:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        mime_type:
          type: string
        size_bytes:
          type: integer
        size_formatted:
          type: string
        area_id:
          type: integer
        uploaded_by:
          $ref: '#/components/schemas/User'
        status:
          type: boolean
        created_at:
          type: string
          format: date-time
    FileListResponse:
      type: object
      properties:
        files:
          type: array
          items:
            $ref: '#/components/schemas/File'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    Share:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        external_email:
          type: string
          format: email
        status:
          type: string
          enum: [pendente, ativo, aprovado, rejeitado, concluido, expirado, cancelado]
        expiration_hours:
          type: integer
        expires_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        files_count:
          type: integer
    ShareDetails:
      allOf:
        - $ref: '#/components/schemas/Share'
        - type: object
          properties:
            created_by:
              $ref: '#/components/schemas/User'
            files:
              type: array
              items:
                $ref: '#/components/schemas/File'
            approval:
              type: object
              nullable: true
              properties:
                approved_by:
                  $ref: '#/components/schemas/User'
                approved_at:
                  type: string
                  format: date-time
                comments:
                  type: string
            rejection:
              type: object
              nullable: true
              properties:
                rejected_by:
                  $ref: '#/components/schemas/User'
                rejected_at:
                  type: string
                  format: date-time
                reason:
                  type: string
    CreateShareRequest:
      type: object
      required: [external_email, file_ids]
      properties:
        name:
          type: string
        description:
          type: string
        external_email:
          type: string
          format: email
        file_ids:
          type: array
          items:
            type: integer
          minItems: 1
        expiration_hours:
          type: integer
          default: 72
          minimum: 1
          maximum: 168
        area_id:
          type: integer
    ShareListResponse:
      type: object
      properties:
        shares:
          type: array
          items:
            $ref: '#/components/schemas/Share'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # Supervisor Schemas
...
    Pagination:
      type: object
      properties:
        current_page:
          type: integer
        total_pages:
          type: integer
        total_items:
          type: integer
        items_per_page:
          type: integer